Can someone make a plugin for Minecraft 1.20.2?
I'm trying to develop a plugin for my server, but I'm having trouble. It seems that the program isn't saving the plugin correctly and is showing programming errors. The plugin in question is a jujutsu kaisen system in Minecraft. Could someone help me create this from scratch? I'll make the plugin available for free for everyone to download on SpringMC.